dashboard/backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
numpy>=1.26
pandas>=2.0
catboost>=1.2
holidays>=0.40
google-generativeai>=0.4
fpdf2>=2.7
openpyxl>=3.1
python-dotenv>=1.0
